fix type warnings in random.c

+++ src/lib/random.c
@@ -10,15 +10,15 @@
 #define BITS_PER_ROLL ((INT_SIZE*8)/ROLLS_PER_INT)
 
 unsigned char die_buffer[DIE_BUFFER_SIZE];
-int die_loc = DIE_BUFFER_SIZE;
+size_t die_loc = DIE_BUFFER_SIZE;
 
 int roll_die(int sides) {
 
   if(sides < MAX_BUFFER_ROLL) {
     if(die_loc >= DIE_BUFFER_SIZE) {
-        for(int k = 0; k < DIE_BUFFER_SIZE; k += ROLLS_PER_INT) {
+        for(size_t k = 0; k < DIE_BUFFER_SIZE; k += ROLLS_PER_INT) {
           int num = rand();
-          for(int i = 0; i < ROLLS_PER_INT; i++) {
+          for(size_t i = 0; i < ROLLS_PER_INT; i++) {
             die_buffer[k + i] = num & MAX_BUFFER_ROLL;
             num >>= BITS_PER_ROLL;
           }
